feat: Sprints 47-49 — four Tier-1 commercial frontiers
R1 15-State ESKF:
  - Closed-loop quaternion attitude feedback with online accel/gyro bias
  - Backward RTS smoother, coupled NHC/ZUPT constraints
  - Odaiba benchmark: p50=2.309m (target <2.50m), RMS=4.642m (target <5.20m)

R2 Integer PPP-AR:
  - LAMBDA integer resolution with discrimination ratios up to 1.9e9
  - SINEX BIA parser, AR-handler with thread_local epoch tracking
  - F9P kinematic: p50=0.679m, RMS=0.680m (sub-meter, GPS-only)
  - WTZR static: p50=0.841m, RMS=0.834m (sub-meter)

R3 VRS Network RTK:
  - Delaunay triangulation, multi-baseline DD network adjustment
  - VRS observation synthesis with ionosphere/troposphere interpolation
  - Network fused p50=0.024m, fix rates 92-94%

R4 Composite TC-PPP/INS and TC-RTK/INS:
  - Tightly-coupled integration of ESKF with PPP and RTK engines
  - 205 E2E requirement-driven tests across Tiers 1-4

Bugfixes:
  - Fixed d_J2000 astronomical century offset in tides.rs (4.5-day error)
  - Refactored global Mutex trackers to thread_local in ar_handler/epoch
  - Fixed CSRS position parser for F9P benchmark comparison

86 files changed, +11179/-1459 lines

+++ crates/gneiss-rtk/src/swfg/imu_preintegration/tests.rs
@@ -115,6 +115,6 @@ use std::collections::BTreeMap;
 
     #[test]
     fn time_diff_wraps_correctly() {
-        let dt = time_diff_us(u32::MAX - 1000, 1000);
+        let dt = time_diff_us(u64::MAX - 1000, 1000);
         assert!((dt - 2001.0 / 1_000_000.0).abs() < 1e-9);
     }
